Revert settings
